first push and devlog 20250206
This devlog focuses on estimating the refugia of origin following a scenario where two expansions waves collide. Brownian motion on a tree is used to model movement in the recent past and the probability distribution of ancestral location at the time that the waves meet is used to estimate which refugia the lineage is from. Further details can be found in `devlog/20250206/20250206.md`